Add plugin roll
