+++ config/command.yml
@@ -46,7 +46,7 @@ gacha:
   - ^十连 *$
   - ^卡池 +(常驻|角色|武器)
 tools:
-  - ^.(r[0-9]+)?d[0-9]+$
+  - ^roll *([+-]?[0-9]+)?$
   - ^help *$
   - ^#repo *$
   - ^#fb +.*
