+++ src/plugins/tools/help.js
@@ -14,7 +14,8 @@ const helpMessage =
 12. 武器                     -> 查询武器强化信息
 13. 天赋                     -> 查询天赋素材信息
 14. 周本                     -> 查询周本材料信息
-15. 管理                     -> 查询管理命令
+15. roll   [数字]            -> 掷骰子
+16. 管理                     -> 查询管理命令
 
 参数：
 1. <> 表示必填，[] 表示可选，前面需加空格
